server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
numpy>=1.26.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
openai>=1.30.0
pyyaml>=6.0.1
websockets>=12.0
openenv-core>=0.1.0
